+++ src/cas.rs
@@ -1114,6 +1114,7 @@ fn modifies_operable(op: &Op, space: SpaceId, storage: &Catalog) -> bool {
         Ddl::DropTable { id, .. } => *id == space,
         Ddl::TruncateTable { id, .. } => *id == space,
         Ddl::ChangeFormat { table_id, .. } => *table_id == space,
+        Ddl::RenameTable { table_id, .. } => *table_id == space,
         Ddl::CreateIndex { .. } => false,
         Ddl::DropIndex { .. } => false,
         Ddl::CreateProcedure { .. } => false,